+++ src/containers/auth/Login.tsx
@@ -1,13 +1,8 @@
 import React from 'react';
 import styled from '@emotion/native';
-import AppleLogo from '../../assets/icons/logo/apple.svg';
-import KakaoLogo from '../../assets/icons/logo/kakao.svg';
-import {
-  Alert,
-  Text,
-  TouchableOpacity,
-  TouchableWithoutFeedbackProps,
-} from 'react-native';
+import { Alert, Text } from 'react-native';
+import LoginButton from '../../components/LoginButton';
+import TextLink from '../../components/TextLink';
 
 const LoginContainer = () => {
   return (
@@ -48,57 +43,3 @@ export default LoginContainer;
 const LoginButtonWrapper = styled.View`
   margin: 16px 37px;
 `;
-
-interface ILoginButtonProps extends TouchableWithoutFeedbackProps {
-  type: 'apple' | 'kakao';
-}
-
-const LoginButton: React.FC<ILoginButtonProps> = ({ type, onPress }) => {
-  return (
-    <TouchableOpacity
-      activeOpacity={0.8}
-      onPress={onPress}
-      style={{
-        height: 54,
-        marginBottom: 8,
-        borderRadius: 8,
-        backgroundColor: `${type === 'apple' ? '#000' : '#FFE812'}`,
-        justifyContent: 'center',
-        alignItems: 'center',
-        flexDirection: 'row',
-      }}>
-      {type === 'apple' ? (
-        <AppleLogo width={24} height={24} />
-      ) : (
-        <KakaoLogo width={24} height={24} />
-      )}
-      <Text
-        numberOfLines={1}
-        adjustsFontSizeToFit
-        style={{
-          color: `${type === 'apple' ? '#fff' : '#000'}`,
-          fontWeight: '600',
-          fontSize: 16,
-          letterSpacing: -0.6,
-          marginLeft: 6,
-        }}>
-        {type === 'apple' ? '애플' : '카카오'}로 시작하기
-      </Text>
-    </TouchableOpacity>
-  );
-};
-
-const TextLink: React.FC<TouchableWithoutFeedbackProps> = ({
-  onPress,
-  children,
-}) => {
-  return (
-    <Text
-      onPress={onPress}
-      numberOfLines={1}
-      adjustsFontSizeToFit
-      style={{ textDecorationLine: 'underline' }}>
-      {children}
-    </Text>
-  );
-};
